+++ public/pages/profile.html
@@ -6,6 +6,11 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Book Swap</title>
     <script src="https://kit.fontawesome.com/f15fba9ca3.js" crossorigin="anonymous"></script>
+    <link rel="stylesheet" href='../style/profile.css'>
+    <link
+    rel="icon"
+    href="https://th.bing.com/th/id/OIP.r-UF3h84JiIwOhrUjSSe3QHaE8?w=261&h=180&c=7&o=5&pid=1.7"
+    />
 </head>
 <body>
     <header>
@@ -18,10 +23,12 @@
         </nav>
     </header>
     <section class="user-info-section">
+        <div class='user'>
         <h2>Ahmad Sabbah</h2>
         <div class="avatar">
             <img src="../images/Profile_avatar_placeholder_large.png" alt="avatar" class="avatar-img">
         </div>
+    </div>
         <button>Add a Book</button>
     </section>
     <section class="added-books">
