+++ public/pages/signUp.html
@@ -13,9 +13,6 @@
   </head>
   <body>
     <div class="container">
-      <header>
-        <h1 class="logo">Book Swap</h1>
-      </header>
       <section class="login-section">
         <form action="/login" method="post">
           <h1 class="login-title">Sign up </h1>
