+++ public/style/books.css
@@ -1,0 +1,6 @@
+* {
+  padding: 0;
+  margin: 0;
+  outline: none;
+  box-sizing: border-box;
+}
