+++ public/style/style.css
@@ -5,6 +5,12 @@
   box-sizing: 0;
 
 }
+body{
+  background-image: url('https://images.unsplash.com/photo-1536965764833-5971e0abed7c?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=750&q=80');
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-position:top;
+}
 .container {
   width: 70%;
   display: flex;
@@ -22,18 +28,14 @@
   align-items: center;
   margin: 15px;
 }
-.logo {
-  margin-bottom: 10px;
-  color: #820682;
 
-}
 form {
   margin-top: 30px;
   display: flex;
   align-items: center;
   justify-content: center;
   flex-direction: column;
-  padding: 10px;
+  padding: 14px;
   background-color: #F6E7E1;
   font-family:Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;
   border-radius: 50px ;
