+++ src/server.js
@@ -1,0 +1,6 @@
+const app = require('./app');
+
+app.listen(3000 , () => {
+    console.log('Server is working! Yahooooooooo!');
+} );
+
